requirements.txt
apimatic-core==0.2.18
apimatic-core-interfaces==0.1.6
apimatic-requests-client-adapter==0.1.7
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
decouple==0.0.7
deprecation==2.1.0
Django==5.1.4
django-allauth==65.3.1
django-countries==7.6.1
django-crispy-forms==2.3
django-paypal==2.1
django-widget-tweaks==1.5.0
idna==3.10
jsonpickle==3.3.0
jsonpointer==2.4
msgpack==1.1.0
packaging==24.2
phonenumbers==8.13.54
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
square==0.0.3
squareup==40.0.0.220250123
stripe==11.4.1
typing_extensions==4.12.2
urllib3==2.3.0
